+++ code/test/sixsq/nuvla/server/resources/customer_lifecycle_test.clj
@@ -1,6 +1,7 @@
 (ns sixsq.nuvla.server.resources.customer-lifecycle-test
   (:require
     [clojure.data.json :as json]
+    [clojure.string :as str]
     [clojure.test :refer [deftest is use-fixtures]]
     [clojure.tools.logging :as log]
     [environ.core :as env]
@@ -23,15 +24,6 @@
 (def base-uri (str p/service-context t/resource-type))
 
 
-(def valid-acl {:owners ["group/nuvla-admin"]})
-
-
-(def timestamp "1964-08-25T10:00:00.00Z")
-
-
-(def test-identifier "some-user-identifer")
-
-
 (def valid-entry {:fullname     "toto"
                   :address      {:street-address "Av. quelque chose"
                                  :city           "Meyrin"
@@ -40,7 +32,6 @@
                   :subscription {:plan-id       "price_1GzO4WHG9PNMTNBOSfypKuEa"
                                  :plan-item-ids ["price_1GzO8HHG9PNMTNBOWuXQm9zZ"
                                                  "price_1GzOC6HG9PNMTNBOEb5819lm"
-                                                 "price_1GzOdmHG9PNMTNBOCvJLV4pT"
                                                  "price_1GzOfLHG9PNMTNBO0l2yDtPS"]}})
 
 
@@ -158,7 +149,11 @@
                                       (ltu/is-operation-present :upcoming-invoice)
                                       (ltu/is-operation-present :list-invoices)
                                       (ltu/is-operation-present :add-coupon)
-                                      (ltu/is-operation-present :remove-coupon))
+                                      (ltu/is-operation-present :remove-coupon)
+                                      (ltu/is-key-value
+                                        #(str/starts-with? % "cus_") :customer-id true)
+                                      (ltu/is-key-value
+                                        #(str/starts-with? % "sub_") :subscription-id true))
               create-setup-intent (ltu/get-op-url customer-response :create-setup-intent)
               add-coupon          (ltu/get-op-url customer-response :add-coupon)]